Extract S3 and Kafka stages into per-source files
