+++ crates/clickhouse-cloud-api/tests/integration/stages/kinesis.rs
@@ -1,0 +1,8 @@
+//! Kinesis ClickPipe stage: per-test stream + IAM_ROLE auth.
+//!
+//! Stub for subagent to fill in. Expected entry point:
+//!
+//!     pub async fn run_kinesis_stage(sctx: StageCtx<'_>) -> StageOutcome
+//!
+//! See `stages/s3.rs` for the closest template (IAM_ROLE pattern + per-service
+//! trust policy).
