+++ crates/clickhouse-cloud-api/tests/integration/stages/mysql.rs
@@ -1,0 +1,8 @@
+//! MySQL ClickPipe stage: per-test EC2 + Basic auth + CDC.
+//!
+//! Stub for subagent to fill in. Expected entry point:
+//!
+//!     pub async fn run_mysql_stage(sctx: StageCtx<'_>) -> StageOutcome
+//!
+//! See `stages/kafka.rs` for the closest template (per-test EC2 + EIP +
+//! user_data + cert generation pattern).
